requirements.txt
aenum==3.1.15
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
attrs==23.2.0
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.0.7
cycler==0.11.0
Deprecated==1.2.14
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
fonttools==4.38.0
frozenlist==1.4.1
future==1.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.4
line-bot-sdk==3.11.0
Mako==1.3.5
MarkupSafe==2.1.5
matplotlib==3.7.0
multidict==6.0.5
numpy==1.24.2
packaging==23.0
Pillow==9.4.0
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.31.0
scipy==1.10.1
six==1.16.0
SQLAlchemy==2.0.31
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
wrapt==1.16.0
yarl==1.9.4
